+++ server/controllers/taxController.js
@@ -2,7 +2,7 @@ import axios from 'axios';
 
 import { getHeaders } from '../utilities/requestUtilities';
 import logger from '../utilities/middleware/logger';
-import { CATEGORY_ENDPOINT } from '../utilities/endpoints';
+import { TAX_ENDPOINT } from '../utilities/endpoints';
 
 
 export async function get(request, response) {
@@ -10,31 +10,31 @@ export async function get(request, response) {
   if (request.params.id && isNaN(request.params.id)) {
     return response
       .status(400)
-      .send('The category {id} must be a number representing the categoryID');
+      .send('The tax {id} must be a number representing the taxID');
   }
   try {
     const token = request.headers.authorization;
     const headers = getHeaders(token);
 
-    const categoryEndpoint = request.params.id
-      ? `${CATEGORY_ENDPOINT}/${request.params.id}`
-      : CATEGORY_ENDPOINT;
+    const taxEndpoint = request.params.id
+      ? `${TAX_ENDPOINT}/${request.params.id}`
+      : TAX_ENDPOINT;
 
-    const categoryResponse = await axios.get(categoryEndpoint, headers);
+    const taxResponse = await axios.get(taxEndpoint, headers);
 
-    if (!categoryResponse.data.data) {
+    if (!taxResponse.data.data) {
 
-      logger.info(`Invalid request: token: ${token}, endpoint: ${categoryEndpoint}`);
+      logger.info(`Invalid request: token: ${token}, endpoint: ${taxEndpoint}`);
       return response
         .status(400)
         .send('Bad request');
     }
 
-    const categoriesData = categoryResponse.data.data;
+    const taxData = taxResponse.data.data;
 
     return response
       .status(200)
-      .json(categoriesData);
+      .json(taxData);
 
   } catch (error) {
 
@@ -45,22 +45,22 @@ export async function get(request, response) {
   }
 }
 
-export async function deleteCategory(request, response) {
+export async function deleteTax(request, response) {
 
-  const categoryId = request.params.id;
+  const taxId = request.params.id;
 
-  if (!categoryId) {
+  if (!taxId) {
     return response
       .status(400)
-      .send('This resource expects a category id');
+      .send('This resource expects a tax id');
   }
 
   try {
     const token = request.headers.authorization;
     const headers = getHeaders(token);
-    const categoryEndpoint = `${CATEGORY_ENDPOINT}/${request.params.id}`;
+    const taxEndpoint = `${TAX_ENDPOINT}/${request.params.id}`;
 
-    await axios.delete(categoryEndpoint, headers);
+    await axios.delete(taxEndpoint, headers);
 
     return response
       .status(200)
@@ -69,90 +69,90 @@ export async function deleteCategory(request, response) {
   } catch (error) {
 
     logger.info(error);
-    logger.info(`Error deleting: ${categoryId}`);
+    logger.info(`Error deleting: ${taxId}`);
     return response
       .status(400)
-      .send('Failed delete the category');
+      .send('Failed delete the tax item');
   }
 }
 
 export async function put(request, response) {
 
-  const categoryId = request.params.id;
-  const categoryBody = request.body;
+  const taxId = request.params.id;
+  const taxBody = request.body;
 
   if (request.params.id && isNaN(request.params.id)) {
     return response
       .status(400)
-      .send('The category {id} must be a number representing the categoryId');
+      .send('The tax {id} must be a number representing the taxID');
   }
 
-  if (!categoryBody) {
+  if (!taxBody) {
     logger.error(`The request {body} cannot be null, ${request.originalUrl}`);
     return response
       .status(400)
-      .send('The category {body} cannot be empty');
+      .send('The tax {body} cannot be empty');
   }
 
   try {
 
     const token = request.headers.authorization;
     const headers = getHeaders(token);
-    const endpoint = `${CATEGORY_ENDPOINT}/${categoryId}`;
+    const endpoint = `${TAX_ENDPOINT}/${taxId}`;
 
-    const postedResponse = await axios.put(endpoint, categoryBody, headers);
-    const updatedCategory = postedResponse.data;
+    const postedResponse = await axios.put(endpoint, taxBody, headers);
+    const updatedTax = postedResponse.data;
 
     return response
       .status(200)
-      .json(updatedCategory);
+      .json(updatedTax);
 
   } catch (error) {
 
     logger.info(error);
-    logger.info(`Error updating: ${categoryId}`);
-    logger.info(`Error updating: ${JSON.stringify(categoryBody)}`);
+    logger.info(`Error updating: ${taxId}`);
+    logger.info(`Error updating: ${JSON.stringify(taxBody)}`);
     return response
       .status(400)
-      .send('Failed to update the category');
+      .send('Failed to update the tax item');
   }
 
 }
 
 export async function post(request, response) {
 
-  const categoryBody = request.body;
+  const taxBody = request.body;
   if (request.params.id) {
     return response
       .status(400)
       .send('This resource does not accept an id');
   }
 
-  if (!categoryBody) {
+  if (!taxBody) {
     logger.error(`The request {body} cannot be null, ${request.originalUrl}`);
     return response
       .status(400)
-      .send('The category {body} cannot be empty');
+      .send('The tax {body} cannot be empty');
   }
 
   try {
     const token = request.headers.authorization;
     const headers = getHeaders(token);
 
-    const postedResponse = await axios.post(CATEGORY_ENDPOINT, categoryBody, headers);
-    const newCategory = postedResponse.data;
+    const postedResponse = await axios.post(TAX_ENDPOINT, taxBody, headers);
+    const newTax = postedResponse.data;
 
     return response
       .status(200)
-      .json(newCategory);
+      .json(newTax);
 
   } catch (error) {
 
     logger.info(error);
-    logger.info(`Error posting a new category: ${JSON.stringify(categoryBody)}`);
+    logger.info(`Error posting a new tax: ${JSON.stringify(taxBody)}`);
     return response
       .status(400)
-      .send('Failed to create a new category');
+      .send('Failed to create a new tax');
   }
 }
 
@@ -160,5 +160,5 @@ export default {
   get,
   put,
   post,
-  deleteCategory
+  deleteTax
 };