+++ server/utilities/endpoints.js
@@ -6,3 +6,4 @@ export const ITEM_ENDPOINT = `${DOMAIN_ENDPOINT}/api/dashboard/item`;
 export const LOGIN_TOKEN_ENDPOINT = `${DOMAIN_ENDPOINT}/api/dashboard/security/token`;
 export const ITEM_IMAGE_ENDPOINT = `${DOMAIN_ENDPOINT}/api/dashboard/image`;
 export const CATEGORY_ENDPOINT = `${DOMAIN_ENDPOINT}/api/dashboard/itemcategories`;
+export const TAX_ENDPOINT = `${DOMAIN_ENDPOINT}/api/dashboard/taxes`;